+++ workspaces/lightspeed/plugins/lightspeed/src/translations/ref.ts
@@ -79,13 +79,12 @@ export const lightspeedMessages = {
 
   // Disclaimers
   'disclaimer.withValidation':
-    "Developer Lightspeed can answer questions on many topics using your configured models. Developer Lightspeed's responses are influenced by the Red Hat Developer Hub documentation but Developer Lightspeed does not have access to your Software Catalog, TechDocs, or Templates etc. Developer Lightspeed uses question (prompt) validation to ensure that conversations remain focused on technical topics relevant to Red Hat Developer Hub, such as Backstage, Kubernetes, and OpenShift. Do not include personal or sensitive information in your input. Interactions with Developer Lightspeed may be reviewed and used to improve products or services.",
+    "This feature uses AI technology. Do not include any personal information or any other sensitive information in your input. Interactions may be used to improve Red Hat's products or services.",
   'disclaimer.withoutValidation':
-    "Developer Lightspeed can answer questions on many topics using your configured models. Developer Lightspeed's responses are influenced by the Red Hat Developer Hub documentation but Developer Lightspeed does not have access to your Software Catalog, TechDocs, or Templates etc. Do not include personal or sensitive information in your input. Interactions with Developer Lightspeed may be reviewed and used to improve products or services.",
+    "This feature uses AI technology. Do not include any personal information or any other sensitive information in your input. Interactions may be used to improve Red Hat's products or services.",
 
   // Footer and feedback
-  'footer.accuracy.label':
-    'Always check AI/LLM generated responses for accuracy prior to use.',
+  'footer.accuracy.label': 'Always review AI generated content prior to use.',
   'footer.accuracy.popover.title': 'Verify accuracy',
   'footer.accuracy.popover.description':
     "While Developer Lightspeed strives for accuracy, there's always a possibility of errors. It's a good practice to verify critical information from reliable sources, especially if it's crucial for decision-making or actions.",