+++ workspaces/adoption-insights/plugins/analytics-module-adoption-insights/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @red-hat-developer-hub/backstage-plugin-analytics-module-adoption-insights
 
+## 0.5.0
+
+### Minor Changes
+
+- 44dd4f2: Backstage version bump to v1.44.2
+
+## 0.4.1
+
 ## 0.3.0
 
 ### Minor Changes