+++ workspaces/ai-integrations/packages/app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # app
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [44c967a]
+  - @red-hat-developer-hub/backstage-plugin-ai-experience@0.4.0
+
 ## 0.0.3
 
 ### Patch Changes